Updated to accommodate new envars from NML. New envar to allow rsync from NCI Host.

+++ lxs599/umplot/tools/mvname.sh
@@ -1,9 +1,9 @@
 #!/bin/csh 
 
 set a=a
-set mvf=saabb # RUNID 1
-set mvt=saaba # RUNID 2
-set ext=pf    # if only one file extension
+set mvf=$1   #saabb # RUNID 1
+set mvt=$2   #saaba # RUNID 2
+set ext=pf   # if only one file extension
 
 set filelist=`ls $mvf$a.p*`
 #set filelist=`ls $mvf$a.$ext*`
@@ -11,8 +11,10 @@ set filelist=`ls $mvf$a.p*`
 
 foreach pfile ( $filelist )
  set newname=`echo $pfile | sed -e s/$mvf/$mvt/`
- echo "Moving " $pfile " to " $newname
- mv $pfile $newname
+ if ( ! -e $newname ) then
+  echo "Moving " $pfile " to " $newname
+  mv $pfile $newname
+ endif
 end
 
 exit
